# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AbortResponse": {
        "additionalProperties": false,
        "description": "Response aborting the run at this check-in.\n\nArgs:\n    kind: Discriminator.",
        "properties": {
          "kind": {
            "const": "abort",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "AbortResponse",
        "type": "object"
      },
...
      "CheckInOption": {
        "description": "One server-supplied option offered at a check-in.\n\nArgs:\n    id: Canonical or authored option id.\n    label: Short user-visible option label.\n    description: Longer user-visible option description.\n    requires_user_input: Whether picking this option requires an\n        additional `params` fill-in on the response.\n    suggested: Whether the server highlights this as the suggested pick.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "requires_user_input": {
            "title": "Requires User Input",
            "type": "boolean"
          },
          "suggested": {
            "default": false,
            "title": "Suggested",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "label",
          "description",
          "requires_user_input"
        ],
        "title": "CheckInOption",
        "type": "object"
      },
      "CheckInOut": {
        "description": "A pending or decided steering check-in.\n\nArgs:\n    check_in_id: The pause event id.\n    kind: Check-in kind.\n    boundary: Which side of the component boundary the pause sits at.\n    component: Composed-chain component the pause concerns, or `None`.\n    stage: Presentation stage key for the pause, or `None`.\n    render: The deterministic render \u2014 content of record; there is no\n        LLM prose wrap.\n    options: Server-supplied options. The client must never invent\n        options.\n    triggers: Fired floor triggers, if any.\n    segment_reentry_allowed: Whether an additive segment re-entry is\n        available at this boundary.\n    rerun_component: Component a replacement re-run would target, or\n        `None` if unavailable.\n    status: Whether the check-in is still pending or already decided.\n    created_at: When the pause was recorded.\n    sequence: The pause event's `event_log` sequence.",
        "properties": {
          "boundary": {
            "enum": [
              "after_component",
              "before_component"
            ],
            "title": "Boundary",
            "type": "string"
          },
          "check_in_id": {
            "format": "uuid",
            "title": "Check In Id",
            "type": "string"
          },
          "component": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Component"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/CheckInOption"
            },
            "title": "Options",
            "type": "array"
          },
          "render": {
            "title": "Render",
            "type": "string"
          },
          "rerun_component": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Rerun Component"
          },
          "segment_reentry_allowed": {
            "title": "Segment Reentry Allowed",
            "type": "boolean"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Stage"
          },
          "status": {
            "enum": [
              "pending",
              "decided"
            ],
            "title": "Status",
            "type": "string"
          },
          "triggers": {
            "items": {
              "$ref": "#/components/schemas/CheckInTrigger"
            },
            "title": "Triggers",
            "type": "array"
          }
        },
        "required": [
          "check_in_id",
          "kind",
          "boundary",
          "render",
          "segment_reentry_allowed",
          "status",
          "created_at",
          "sequence"
        ],
        "title": "CheckInOut",
        "type": "object"
      },
...
      "CheckInTrigger": {
        "description": "One fired floor trigger reported alongside a check-in.\n\nArgs:\n    trigger: Trigger name (e.g. `excluded_large_stratum`).\n    detail: Trigger-specific detail payload.",
        "properties": {
          "detail": {
            "default": null,
            "title": "Detail"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "trigger"
        ],
        "title": "CheckInTrigger",
        "type": "object"
      },
...
      "FreeTextCompileOut": {
        "description": "The 202 body returned for a `free_text` check-in response.\n\nArgs:\n    render: The compiled fan-out render \u2014 the confirm-gate content of\n        record. Nothing applies until confirmed.\n    confirm_token: Opaque token the client echoes back in a\n        `free_text_confirm` response.",
        "properties": {
          "confirm_token": {
            "title": "Confirm Token",
            "type": "string"
          },
          "render": {
            "title": "Render",
            "type": "string"
          }
        },
        "required": [
          "render",
          "confirm_token"
        ],
        "title": "FreeTextCompileOut",
        "type": "object"
      },
      "FreeTextConfirmResponse": {
        "additionalProperties": false,
        "description": "Response confirming or discarding a compiled free-text fan-out.\n\nArgs:\n    kind: Discriminator.\n    confirm_token: Token threading this confirm back to its compile\n        (`FreeTextCompileOut.confirm_token`).\n    apply: `True` to apply the compiled deltas, `False` to discard them.",
        "properties": {
          "apply": {
            "title": "Apply",
            "type": "boolean"
          },
          "confirm_token": {
            "title": "Confirm Token",
            "type": "string"
          },
          "kind": {
            "const": "free_text_confirm",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "confirm_token",
          "apply"
        ],
        "title": "FreeTextConfirmResponse",
        "type": "object"
      },
      "FreeTextResponse": {
        "additionalProperties": false,
        "description": "Response supplying free text for the router to compile.\n\nArgs:\n    kind: Discriminator.\n    text: The user's free-text steer.",
        "properties": {
          "kind": {
            "const": "free_text",
            "title": "Kind",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "text"
        ],
        "title": "FreeTextResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LatestRun": {
        "description": "The derived latest-run read model carried on a project.\n\nArgs:\n    capability_run_id: Identity of the project's most recent run.\n    status: That run's current status.\n    started_at: When that run started executing.\n    ended_at: When that run reached a terminal status, or `None` while\n        still running or paused.",
        "properties": {
          "capability_run_id": {
            "format": "uuid",
            "title": "Capability Run Id",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "paused",
              "succeeded",
              "degraded",
              "failed",
              "aborted",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "capability_run_id",
          "status",
          "started_at"
        ],
        "title": "LatestRun",
        "type": "object"
      },
      "OptionResponse": {
        "additionalProperties": false,
        "description": "Response picking a canonical or authored option.\n\nArgs:\n    kind: Discriminator.\n    option_id: The picked option's id.\n    params: Fill-in for `requires_user_input` options. The spec does\n        not name per-option param shapes, so this stays a constrained\n        free-form mapping rather than a per-option union.",
        "properties": {
          "kind": {
            "const": "option",
            "title": "Kind",
            "type": "string"
          },
          "option_id": {
            "title": "Option Id",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Params"
          }
        },
        "required": [
          "kind",
          "option_id"
        ],
        "title": "OptionResponse",
        "type": "object"
      },
      "PageMeta": {
        "description": "Pagination metadata attached to a paginated list response.\n\nArgs:\n    page: 1-indexed page number returned.\n    page_size: Number of items requested per page (server-capped at\n        `PAGE_SIZE_MAX`).\n    total_items: Total number of items across all pages.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_items": {
            "minimum": 0.0,
            "title": "Total Items",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PageMeta",
        "type": "object"
      },
      "Page_CheckInOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CheckInOut"
            },
            "title": "Data",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "Page[CheckInOut]",
        "type": "object"
      },
...
      "Page_ProjectOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProjectOut"
            },
            "title": "Data",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "Page[ProjectOut]",
        "type": "object"
      },
...
      "ProjectCreate": {
        "additionalProperties": false,
        "description": "Inbound body for `POST /api/v1/projects`.\n\nArgs:\n    name: Project display name, 1-200 characters. Outer whitespace is\n        stripped before the length constraint is applied\n        (`str_strip_whitespace`).\n    question: Optional initial evidence question.",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectOut": {
        "description": "A project resource.\n\nArgs:\n    project_id: The project's identity.\n    name: Current display name.\n    question: Current evidence question, or `None` if not yet set.\n    status: Lifecycle status (`active` unless archived; no hard delete).\n    created_at: When the project was created.\n    updated_at: When the project row was last written.\n    archived_at: When the project was archived, or `None` if active.\n    latest_run: The derived latest-run read model, or `None` before any\n        run has been created.",
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "latest_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Policy Atlas API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/projects": {
      "get": {
        "description": "List the authenticated user's projects with derived latest-run state.",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "archived",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "description": "Create one active project owned by the authenticated subject.",
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/check-ins": {
      "get": {
        "description": "Return a latest pending card or the durable steering history projection.\n\nPaginated (rubric item 17 \u2014 check-ins accumulate over a project's life);\nthe pending view is at most one card by construction.",
        "operationId": "list_check_ins_api_v1_projects__project_id__check_ins_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "enum": [
                "pending",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CheckInOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Check Ins",
        "tags": [
          "check-ins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/check-ins/{check_in_id}/response": {
      "post": {
        "description": "Compile or durably answer one check-in, dispatching continuations after commit.",
        "operationId": "respond_to_check_in_api_v1_projects__project_id__check_ins__check_in_id__response_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "check_in_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Check In Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "abort": "#/components/schemas/AbortResponse",
                    "free_text": "#/components/schemas/FreeTextResponse",
                    "free_text_confirm": "#/components/schemas/FreeTextConfirmResponse",
                    "option": "#/components/schemas/OptionResponse"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OptionResponse"
                  },
                  {
                    "$ref": "#/components/schemas/FreeTextResponse"
                  },
                  {
                    "$ref": "#/components/schemas/FreeTextConfirmResponse"
                  },
                  {
                    "$ref": "#/components/schemas/AbortResponse"
                  }
                ],
                "title": "Response"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreeTextCompileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Respond To Check In",
        "tags": [
          "check-ins"
        ]
      }
    },
...
  }
}